Added basic eexamples on how to include partials. #master
Had to changes a little bit the current architecture.
Every html page is in views/ directory. Styleguide is under styleguide/ directory
and page_parts didnt changed.

Added some text in the page_parts/header.html and an example on how to include it
in page_parts/pages.html.

The readme will guide you on how to run it.