Adds loading from cdn with local fetch fallback #loading
Using yepnope, synchronously fetch with fallback for both js and css.
http://sitepoint.com/7-reasons-to-use-a-cdn.
But in case, the cdn server is not loading correctly the needed library,
we need to local fetch the library.
An optimization could be to asynchronously load the css and js
with libraries such as (requirejs)[http://requirejs.org]
 and (loadCss)[https://github.com/filamentgroup/loadCSS].
As the project is not huge, the perf impact of sync load is not big deal.